CombineWrapper tests
